# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ErTraining",
        "version": "1.0.0"
    },
    "paths": {
        "/api/courses/coursesList": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Tutor, - ritorna la lista dei corsi e la loro descrizione",
                "description": "Get the list of all courses",
                "operationId": "coursesList",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/courses/getCourse/{id}": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Tutor - ritorna tutte le informazioni di un determinato corso",
                "description": "Get the course by id",
                "operationId": "getCourse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "course Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/courses/getUsersCourse/{id}": {
            "get": {
                "tags": [
                    "UsersCourses"
                ],
                "summary": "Ritorna tutte le informazioni degli utenti iscritti ad un corso",
                "description": "Get all the users of a course",
                "operationId": "getUsersCourse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "course Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/courses/addUsersCourse/{course_id}": {
            "post": {
                "tags": [
                    "UsersCourses"
                ],
                "summary": "Tutor - aggiunge utenti al corso",
                "description": "Add user to course",
                "operationId": "addUsersCourse",
                "parameters": [
                    {
                        "name": "course_id",
                        "in": "path",
                        "description": "course Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "User credentials",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "format": "user_id"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/courses/removeUsersCourse/{course_id}": {
            "delete": {
                "tags": [
                    "UsersCourses"
                ],
                "summary": "Tutor - rimuove un utente dal corso, passandogli l'id dell'utente nel body",
                "description": "remove user from course",
                "operationId": "removeUsersCourse",
                "parameters": [
                    {
                        "name": "course_id",
                        "in": "path",
                        "description": "course id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "User credentials",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "format": "user_id"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/courses/addCourse": {
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "Tutor - Permette di creare un corso",
                "description": "Add course",
                "operationId": "addCourse",
                "requestBody": {
                    "description": "Course credentials",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "state",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "name"
                                    },
                                    "state": {
                                        "type": "string",
                                        "format": "state"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/courses/editCourse/{id}": {
            "put": {
                "tags": [
                    "Courses"
                ],
                "summary": "Tutor - modifica i dati del corso, si possono modificare nome, stato e descrizione",
                "description": "Edit course",
                "operationId": "editCourse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "course Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "User credentials",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "state",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "name"
                                    },
                                    "state": {
                                        "type": "string",
                                        "format": "state"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "description"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/courses/deleteCourse/{id}": {
            "delete": {
                "tags": [
                    "Courses"
                ],
                "summary": "Tutor, cancella il corso indicato",
                "description": "delete the course by id",
                "operationId": "deleteCourse",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "course Id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}